+++ packages/collaboration/comments/src/Comment.ts
@@ -1,0 +1,4 @@
+export interface Comment {
+  id: any;
+  text: string;
+}
