+++ packages/collaboration/comments/src/Thread.ts
@@ -1,0 +1,11 @@
+import { Comment } from './Comment';
+
+export interface Thread {
+  id: any;
+  comments: Comment[];
+  isResolved: boolean;
+}
+
+export function isFirstComment(thread: Thread, comment: Comment): boolean {
+  return thread.comments.indexOf(comment) === 0;
+}
