+++ packages/collaboration/comments/src/index.ts
@@ -3,3 +3,5 @@ export * from './types';
 export * from './transforms/index';
 export * from './findThreadNodeEntries';
 export * from './findSelectedThreadNodeEntry';
+export * from './Thread';
+export * from './Comment';
