+++ packages/collaboration/comments/src/transforms/upsertThread.ts
@@ -1,5 +1,5 @@
 import { PlateEditor } from '@udecode/plate-core';
-import { Thread } from '../types';
+import { Thread } from '../Thread';
 import { upsertThreadAtSelection } from './upsertThreadAtSelection';
 
 export function upsertThread<T = {}>(
