+++ packages/collaboration/comments/src/types.ts
@@ -1,17 +1,8 @@
+import { Thread } from './Thread';
+
 export interface ThreadNodeData {
   thread: Thread;
   selected: boolean;
 }
 
 export interface ThreadPlugin {}
-
-export interface Comment {
-  id: any;
-  text: string;
-}
-
-export interface Thread {
-  id: any;
-  comments: Comment[];
-  isResolved: boolean;
-}
